requirements.txt
streamlit
python-amazon-paapi